Fixes:   * Fixed missing TransactionAlreadyInProgressException   * Fixed incorrect docblocks   * Moved VerifyEventIsAClass to a trait
