+++ src/EventStore/InMemoryEventStore/TransactionAlreadyInProgressException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rawkode\Eidetic\EventStore\InMemoryEventStore;
+
+final class TransactionAlreadyInProgressException extends \Exception
+{
+}
